Encoding ok

+++ pod/dressing/views.py
@@ -9,7 +9,8 @@ from django.contrib import messages
 from django.db.models import Q
 
 from pod.video.models import Video
-from .forms import DressingForm
+from pod.video_encode_transcript.encode import start_encode
+from .forms import DressingForm, DressingDeleteForm
 from .models import Dressing
 
 
@@ -24,14 +25,32 @@ def video_dressing(request, slug):
     user = request.user
     users_groups = user.owner.accessgroup_set.all()
     dressings = Dressing.objects.filter(
-        Q(owners=user) |
-        Q(users=user) |
-        Q(allow_to_groups__in=users_groups)).distinct()
+        Q(owners=user) | Q(users=user) | Q(allow_to_groups__in=users_groups)).distinct()
+
+    if request.method == 'POST':
+        selected_dressing_value = request.POST.get("selected_dressing_value")
+        if selected_dressing_value is not None:
+            selected_dressing = get_object_or_404(Dressing, pk=selected_dressing_value)
+        existing = Dressing.objects.filter(videos=video)
+        for dressing in existing:
+            dressing.videos.remove(video)
+            dressing.save()
+        if selected_dressing_value is not None:
+            selected_dressing.videos.add(video)
+            selected_dressing.save()
+            start_encode(video.id)
+        elif selected_dressing_value is None:
+            start_encode(video.id)
+
+    try:
+        current = Dressing.objects.get(videos=video)
+    except Dressing.DoesNotExist:
+        current = None
 
     return render(
         request,
         "video_dressing.html",
-        {"video": video, "dressings": dressings}
+        {"video": video, "dressings": dressings, "current": current}
     )
 
 
@@ -47,9 +66,14 @@ def dressing_edit(request, dressing_id):
     )
 
     if request.method == 'POST':
-        form_dressing = DressingForm(request.POST, instance=dressing_edit)
+        form_dressing = DressingForm(
+            request.POST,
+            instance=dressing_edit,
+            user=request.user,
+        )
         if form_dressing.is_valid():
-            messages.add_message(request, messages.INFO, _("Valid"))
+            messages.add_message(request, messages.INFO,
+                                 _("The changes have been saved."))
             form_dressing.save()
             return redirect(reverse("dressing:my_dressings"))
 
@@ -59,10 +83,11 @@ def dressing_edit(request, dressing_id):
         {'dressing_edit': dressing_edit, "form": form_dressing})
 
 
+@csrf_protect
 @login_required(redirect_field_name="referrer")
 def dressing_create(request):
     if request.method == 'POST':
-        form_dressing = DressingForm(request.POST)
+        form_dressing = DressingForm(request.POST, user=request.user)
         if form_dressing.is_valid():
             form_dressing.save()
             return redirect('dressing:my_dressings')
@@ -75,16 +100,30 @@ def dressing_create(request):
         {'dressing_create': dressing_create, "form": form_dressing})
 
 
+@csrf_protect
 @login_required(redirect_field_name="referrer")
 def dressing_delete(request, dressing_id):
+    """Delete the dressing identified by 'id'."""
     dressing = get_object_or_404(Dressing, id=dressing_id)
 
-    if request.method == 'POST':
-        dressing.delete()
-        return redirect('dressing:my_dressings')
+    form = DressingDeleteForm()
 
-    return render(request, 'dressing_confirm_delete.html',
-                  {'dressing': dressing})
+    if request.method == 'POST':
+        form = DressingDeleteForm(request.POST)
+        if form.is_valid():
+            dressing.delete()
+            messages.add_message(request, messages.INFO,
+                                 _("The dressing has been deleted."))
+            return redirect('dressing:my_dressings')
+        else:
+            messages.add_message(
+                request,
+                messages.ERROR,
+                _("One or more errors have been found in the form."),
+            )
+
+    return render(request, 'dressing_delete.html',
+                  {'dressing': dressing, "form": form})
 
 
 @csrf_protect
@@ -96,9 +135,7 @@ def my_dressings(request):
     user = request.user
     users_groups = user.owner.accessgroup_set.all()
     dressings = Dressing.objects.filter(
-        Q(owners=user) |
-        Q(users=user) |
-        Q(allow_to_groups__in=users_groups)).distinct()
+        Q(owners=user) | Q(users=user) | Q(allow_to_groups__in=users_groups)).distinct()
 
     return render(request, "my_dressings.html",
                   {'dressings': dressings})
